custom CheckpointManager to save latest and best model
This removes the need for RemoveOldCheckpoints() and other functions
defined in eval_utils.py, so removed the whole file.

delete_excess_models.py is not needed since they will automatically be
deleted by the checkpoint manager.

metrics.py test() needs to return the validation accuracy to pass to the
checkpoint manager.

Previously, you'd have to set model_step as a multiple of log_val_step,
but now the model is saved every model_steps and every log_val_step so
you don't have to worry about it. This means the model won't ever
not exist when testing (like it could have previously).